+++ pyaud/_environ.py
@@ -13,7 +13,7 @@ from environs import Env as _Env
 from ._version import __version__
 
 
-class Environ(_Env):
+class _Environ(_Env):
     """Package's environment variables."""
 
     @property
@@ -182,7 +182,7 @@ class Environ(_Env):
 
 #: package environment, both parsed from .env file (with set defaults
 #: for missing keys), and static values
-environ = Environ()
+environ = _Environ()
 
 
 def initialize_dirs() -> None: