+++ pyaud/_main.py
@@ -5,8 +5,7 @@ pyaud.main
 import logging as _logging
 from pathlib import Path as _Path
 
-from . import _data
-from . import config as _config
+from . import _config, _data
 from . import plugins as _plugins
 from ._cli import Parser as _Parser
 from ._default import register_default_plugins as _register_default_plugins