+++ tests/_test.py
@@ -20,6 +20,9 @@ from subprocess import CalledProcessError
 import pytest
 
 import pyaud
+
+# noinspection PyProtectedMember
+import pyaud._config as pc
 from pyaud import environ as pe
 
 from . import (
@@ -44,6 +47,7 @@ from . import (
     WHITELIST_PY,
     NoColorCapsys,
     Tracker,
+    git,
 )
 
 
@@ -51,9 +55,9 @@ def test_get_branch_unique() -> None:
     """Test that ``get_branch`` returns correct branch."""
     Path(Path.cwd() / README).touch()
     branch = datetime.datetime.now().strftime("%d%m%YT%H%M%S")
-    pyaud.git.add(".", devnull=True)
-    pyaud.git.commit("-m", INITIAL_COMMIT, devnull=True)
-    pyaud.git.checkout("-b", branch, devnull=True)
+    git.add(".", devnull=True)
+    git.commit("-m", INITIAL_COMMIT, devnull=True)
+    git.checkout("-b", branch, devnull=True)
     assert pyaud._utils.branch() == branch  # pylint: disable=protected-access
 
 
@@ -64,10 +68,10 @@ def test_get_branch_initial_commit() -> None:
     commit.
     """
     Path(Path.cwd() / README).touch()
-    pyaud.git.add(".")
-    pyaud.git.commit("-m", INITIAL_COMMIT)
-    pyaud.git.rev_list("--max-parents=0", "HEAD", capture=True)
-    pyaud.git.checkout(pyaud.git.stdout()[0])
+    git.add(".")
+    git.commit("-m", INITIAL_COMMIT)
+    git.rev_list("--max-parents=0", "HEAD", capture=True)
+    git.checkout(git.stdout()[0])
     assert pyaud._utils.branch() is None  # pylint: disable=protected-access
 
 
@@ -90,15 +94,15 @@ def test_loglevel(
         "-vvv": [INFO, DEBUG, DEBUG, DEBUG, DEBUG],
         "-vvvv": [DEBUG, DEBUG, DEBUG, DEBUG, DEBUG],
     }
-    pyaud.config.toml["logging"]["root"]["level"] = default
-    pe.GLOBAL_CONFIG_FILE.write_text(pyaud.config.toml.dumps())
+    pc.toml["logging"]["root"]["level"] = default
+    pe.GLOBAL_CONFIG_FILE.write_text(pc.toml.dumps())
 
     # dummy call to non-existing plugin to evaluate multiple -v
     # arguments
     monkeypatch.setattr(
         PYAUD_PLUGINS_PLUGINS, {"module": lambda *_, **__: None}
     )
-    pyaud.config.configure_global()
+    pc.configure_global()
     main("module", flag)
     assert (
         logging.getLevelName(logging.root.level)
@@ -110,9 +114,7 @@ def test_del_key_in_context():
     """Confirm there is no error raised when deleting temp key-value."""
     obj = {}
     # noinspection PyProtectedMember
-    with pyaud.config.TempEnvVar(  # pylint: disable=protected-access
-        obj, key="value"
-    ):
+    with pc.TempEnvVar(obj, key="value"):  # pylint: disable=protected-access
         assert obj["key"] == "value"
         del obj["key"]
 
@@ -182,15 +184,15 @@ def test_help(
 
 def test_mapping_class() -> None:
     """Get coverage on ``Mapping`` abstract methods."""
-    pyaud.config.toml.clear()
-    assert repr(pyaud.config.toml) == "<_Toml {}>"
-    pyaud.config.toml.update({"key": "value"})
-    assert len(pyaud.config.toml) == 1
-    for key in pyaud.config.toml:
+    pc.toml.clear()
+    assert repr(pc.toml) == "<_Toml {}>"
+    pc.toml.update({"key": "value"})
+    assert len(pc.toml) == 1
+    for key in pc.toml:
         assert key == "key"
 
-    del pyaud.config.toml["key"]
-    assert "key" not in pyaud.config.toml
+    del pc.toml["key"]
+    assert "key" not in pc.toml
 
 
 def test_toml() -> None:
@@ -204,10 +206,8 @@ def test_toml() -> None:
     """
     # base config is created and loaded
     # =================================
-    test_default: t.Dict[t.Any, t.Any] = copy.deepcopy(
-        pyaud.config.DEFAULT_CONFIG
-    )
-    assert dict(pyaud.config.toml) == test_default
+    test_default: t.Dict[t.Any, t.Any] = copy.deepcopy(pc.DEFAULT_CONFIG)
+    assert dict(pc.toml) == test_default
 
     # instantiate a new dict object
     # =============================
@@ -218,9 +218,7 @@ def test_toml() -> None:
         {"class": "logging.handlers.StreamHandler"}
     )
     home_rcfile["logging"]["version"] = 2
-    pe.USER_CONFIG_FILE.write_text(
-        pyaud.config.toml.dumps(home_rcfile), pe.ENCODING
-    )
+    pe.USER_CONFIG_FILE.write_text(pc.toml.dumps(home_rcfile), pe.ENCODING)
 
     # reset the dict to the test default
     # ==================================
@@ -229,13 +227,13 @@ def test_toml() -> None:
     project_rcfile = dict(test_default)
     project_rcfile["logging"]["version"] = 3
     pe.PROJECT_CONFIG_FILE.write_text(
-        pyaud.config.toml.dumps(project_rcfile), pe.ENCODING
+        pc.toml.dumps(project_rcfile), pe.ENCODING
     )
 
     # load "$HOME/.pyaudrc" and then "$PROJECT_DIR/.pyaudrc"
     # ======================================================
     # override "$HOME/.pyaudrc"
-    pyaud.config.load_config()
+    pc.load_config()
     subtotal: t.Dict[str, t.Any] = dict(home_rcfile)
     subtotal["logging"]["version"] = 3
     subtotal["logging"]["handlers"]["default"]["filename"] = str(
@@ -243,7 +241,7 @@ def test_toml() -> None:
             subtotal["logging"]["handlers"]["default"]["filename"]
         ).expanduser()
     )
-    assert dict(pyaud.config.toml) == subtotal
+    assert dict(pc.toml) == subtotal
 
     # load pyproject.toml
     # ===================
@@ -251,30 +249,28 @@ def test_toml() -> None:
     pyproject_dict = {"tool": {pyaud.__name__: test_default}}
     changes = {"clean": {"exclude": []}, "logging": {"version": 4}}
     pyproject_dict["tool"][pyaud.__name__].update(changes)
-    pe.PYPROJECT.write_text(
-        pyaud.config.toml.dumps(pyproject_dict), pe.ENCODING
-    )
+    pe.PYPROJECT.write_text(pc.toml.dumps(pyproject_dict), pe.ENCODING)
 
     # ======================================================
     # override "$HOME/.pyaudrc"
-    pyaud.config.load_config()
+    pc.load_config()
     subtotal["clean"]["exclude"] = []
     subtotal["logging"]["version"] = 4
-    assert dict(pyaud.config.toml) == subtotal
+    assert dict(pc.toml) == subtotal
 
     # load optional rcfile
     # ====================
     # this will override all others when passed to the commandline
     pos = {"audit": {"modules": ["files", "format", "format-docs"]}}
     opt_rc = Path.cwd() / "opt_rc"
-    opt_rc.write_text(pyaud.config.toml.dumps(pos), pe.ENCODING)
+    opt_rc.write_text(pc.toml.dumps(pos), pe.ENCODING)
 
     # load "$HOME/.pyaudrc" and then "$Path.cwd()/.pyaudrc"
     # ======================================================
     # override "$HOME/.pyaudrc"
-    pyaud.config.load_config(opt_rc)
+    pc.load_config(opt_rc)
     subtotal["audit"] = {"modules": ["files", "format", "format-docs"]}
-    assert dict(pyaud.config.toml) == subtotal
+    assert dict(pc.toml) == subtotal
 
 
 def test_toml_no_override_all(monkeypatch: pytest.MonkeyPatch) -> None:
@@ -293,16 +289,15 @@ def test_toml_no_override_all(monkeypatch: pytest.MonkeyPatch) -> None:
     :param monkeypatch: Mock patch environment and attributes.
     """
     monkeypatch.setattr(
-        "pyaud.config.DEFAULT_CONFIG",
-        copy.deepcopy(pyaud.config.DEFAULT_CONFIG),
+        "pyaud._config.DEFAULT_CONFIG",
+        copy.deepcopy(pyaud._config.DEFAULT_CONFIG),
     )
-    pyaud.config.toml.clear()
-    pyaud.config.load_config()  # base key-values
-    pyaud.config.toml.loads(pe.GLOBAL_CONFIG_FILE.read_text(pe.ENCODING))
-    assert dict(pyaud.config.toml) == pyaud.config.DEFAULT_CONFIG
+    pc.toml.clear()
+    pc.load_config()  # base key-values
+    pc.toml.loads(pe.GLOBAL_CONFIG_FILE.read_text(pe.ENCODING))
+    assert dict(pc.toml) == pc.DEFAULT_CONFIG
     pe.USER_CONFIG_FILE.write_text(
-        pyaud.config.toml.dumps({"logging": {"root": {"level": "INFO"}}}),
-        pe.ENCODING,
+        pc.toml.dumps({"logging": {"root": {"level": "INFO"}}}), pe.ENCODING
     )
 
     # should override:
@@ -327,13 +322,13 @@ def test_toml_no_override_all(monkeypatch: pytest.MonkeyPatch) -> None:
     # },
     # and not reduce it to:
     # {"root": {"level": "INFO"}}
-    pyaud.config.load_config()
+    pc.load_config()
 
     # this here would raise a ``ValueError`` if not working as expected,
     # so on its own is an assertion
-    logging_config.dictConfig(pyaud.config.toml["logging"])
-    pyaud.config.DEFAULT_CONFIG["logging"]["root"]["level"] = "INFO"
-    assert dict(pyaud.config.toml) == pyaud.config.DEFAULT_CONFIG
+    logging_config.dictConfig(pc.toml["logging"])
+    pc.DEFAULT_CONFIG["logging"]["root"]["level"] = "INFO"
+    assert dict(pc.toml) == pc.DEFAULT_CONFIG
 
 
 # noinspection DuplicatedCode
@@ -354,21 +349,21 @@ def test_backup_toml() -> None:
     # ==============
     # originally there is no backup file (not until configure_global is
     # run)
-    default_config = dict(pyaud.config.toml)
+    default_config = dict(pc.toml)
     assert not pe.GLOBAL_CONFIG_BAK_FILE.is_file()
 
     # assert corrupt configfile with no backup will simply reset
     configfile_contents = pe.GLOBAL_CONFIG_FILE.read_text(pe.ENCODING)
 
     _corrupt_file(configfile_contents)
-    pyaud.config.configure_global()
-    pyaud.config.toml.loads(pe.GLOBAL_CONFIG_FILE.read_text(pe.ENCODING))
+    pc.configure_global()
+    pc.toml.loads(pe.GLOBAL_CONFIG_FILE.read_text(pe.ENCODING))
 
     # assert corrupt configfile is no same as default
-    assert dict(pyaud.config.toml) == default_config
+    assert dict(pc.toml) == default_config
 
     # create backupfile
-    pyaud.config.configure_global()
+    pc.configure_global()
     assert pe.GLOBAL_CONFIG_BAK_FILE.is_file()
 
     # ensure backupfile is a copy of the original config file
@@ -380,15 +375,15 @@ def test_backup_toml() -> None:
     # change to config
     # ================
     # this setting, by default, is True
-    pyaud.config.toml["logging"]["disable_existing_loggers"] = False
-    pe.GLOBAL_CONFIG_FILE.write_text(pyaud.config.toml.dumps(), pe.ENCODING)
+    pc.toml["logging"]["disable_existing_loggers"] = False
+    pe.GLOBAL_CONFIG_FILE.write_text(pc.toml.dumps(), pe.ENCODING)
 
     # now that there is a change the backup should be different to the
     # original until configure_global is run again
     # read configfile as only that file has been changed
     configfile_contents = pe.GLOBAL_CONFIG_FILE.read_text(pe.ENCODING)
     assert configfile_contents != backupfile_contents
-    pyaud.config.configure_global()
+    pc.configure_global()
 
     # read both, as both have been changed
     configfile_contents = pe.GLOBAL_CONFIG_FILE.read_text(pe.ENCODING)
@@ -407,15 +402,15 @@ def test_backup_toml() -> None:
 
     # resolve corruption
     # ==================
-    pyaud.config.configure_global()
+    pc.configure_global()
     configfile_contents = pe.GLOBAL_CONFIG_FILE.read_text(pe.ENCODING)
     backupfile_contents = pe.GLOBAL_CONFIG_BAK_FILE.read_text(pe.ENCODING)
 
     # configfile should equal the backup file and all changes should be
     # retained
     assert configfile_contents == backupfile_contents
-    pyaud.config.toml.loads(pe.GLOBAL_CONFIG_FILE.read_text(pe.ENCODING))
-    assert pyaud.config.toml["logging"]["disable_existing_loggers"] is False
+    pc.toml.loads(pe.GLOBAL_CONFIG_FILE.read_text(pe.ENCODING))
+    assert pc.toml["logging"]["disable_existing_loggers"] is False
 
 
 def test_register_plugin_name_conflict_error() -> None:
@@ -520,16 +515,16 @@ def test_files_populate_proc(make_tree: t.Any) -> None:
     def _old_files_populate():
         indexed = []
         for path in Path.cwd().rglob("*.py"):
-            if path.name not in pyaud.config.DEFAULT_CONFIG["indexing"][
+            if path.name not in pc.DEFAULT_CONFIG["indexing"][
                 "exclude"
-            ] and not pyaud.git.ls_files(
+            ] and not git.ls_files(
                 "--error-unmatch", path, devnull=True, suppress=True
             ):
                 indexed.append(path)
 
         return indexed
 
-    pyaud.git.add(".")
+    git.add(".")
     start = time.process_time()
     no_commit_files = _old_files_populate()
     stop = time.process_time()
@@ -562,7 +557,7 @@ def test_get_packages(
     # search for only package
     # =======================
     make_tree(Path.cwd(), {"first_package": {INIT: None}})
-    assert pyaud.get_packages() == ["first_package"]
+    assert pyaud._utils.get_packages() == ["first_package"]
     assert pyaud.package() == "first_package"
 
     # search for ambiguous package
@@ -571,7 +566,7 @@ def test_get_packages(
         Path.cwd(),
         {"second_package": {INIT: None}, "third_package": {INIT: None}},
     )
-    assert pyaud.get_packages() == [
+    assert pyaud._utils.get_packages() == [
         "first_package",
         "second_package",
         "third_package",
@@ -581,7 +576,7 @@ def test_get_packages(
     # search for package with the same name as repo
     # =============================================
     make_tree(Path.cwd(), {"repo": {INIT: None}})
-    assert pyaud.get_packages() == [
+    assert pyaud._utils.get_packages() == [
         "first_package",
         "repo",
         "second_package",
@@ -591,7 +586,7 @@ def test_get_packages(
 
     # search for configured package
     # =============================
-    pyaud.config.toml["packages"]["name"] = "second_package"
+    pc.toml["packages"]["name"] = "second_package"
     assert pyaud.package() == "second_package"
 
 
@@ -631,7 +626,7 @@ def test_get_subpackages(
 
     # assert no dot separated packages are returned and that only the
     # parent packages name is returned
-    assert pyaud.get_packages() == ["repo"]
+    assert pyaud._utils.get_packages() == ["repo"]
 
 
 def test_exclude_loads_at_main(main: t.Any) -> None:
@@ -648,21 +643,19 @@ def test_exclude_loads_at_main(main: t.Any) -> None:
         def action(self, *args: t.Any, **kwargs: bool) -> t.Any:
             """Nothing to do."""
 
-    default_config = copy.deepcopy(pyaud.config.DEFAULT_CONFIG)
+    default_config = copy.deepcopy(pc.DEFAULT_CONFIG)
     project_config = copy.deepcopy(default_config)
     project_config["indexing"]["exclude"].append("project")
-    test_project_toml_object = (
-        pyaud.config._Toml()  # pylint: disable=protected-access
-    )
+    test_project_toml_object = pc._Toml()  # pylint: disable=protected-access
     test_project_toml_object.update(project_config)
     pe.PROJECT_CONFIG_FILE.write_text(
         test_project_toml_object.dumps(), pe.ENCODING
     )
-    assert "project" not in pyaud.config.toml["indexing"]["exclude"]
+    assert "project" not in pc.toml["indexing"]["exclude"]
 
     main("plugin")
 
-    assert "project" in pyaud.config.toml["indexing"]["exclude"]
+    assert "project" in pc.toml["indexing"]["exclude"]
 
 
 def test_exclude(make_tree: t.Any) -> None:
@@ -691,7 +684,7 @@ def test_exclude(make_tree: t.Any) -> None:
         },
     )
     exclude = (WHITELIST_PY, "conf.py", "setup.py", "migrations")
-    pyaud.git.add(".")
+    git.add(".")
     pyaud.files.add_exclusions(*exclude)
     pyaud.files.populate()
     assert not any(i in p.parts for i in exclude for p in pyaud.files)
@@ -701,19 +694,15 @@ def test_exclude(make_tree: t.Any) -> None:
 # noinspection DuplicatedCode
 def test_filter_logging_config_kwargs() -> None:
     """Test that no errors are raised for additional config kwargs."""
-    test_default: t.Dict[t.Any, t.Any] = copy.deepcopy(
-        pyaud.config.DEFAULT_CONFIG
-    )
+    test_default: t.Dict[t.Any, t.Any] = copy.deepcopy(pc.DEFAULT_CONFIG)
 
     # patch `DEFAULT_CONFIG` for `TimedRotatingFileHandler`
     logfile = str(Path.cwd() / ".cache" / "pyaud" / "log" / "pyaud.log")
     test_default["logging"]["handlers"]["default"]["filename"] = logfile
     rcfile = dict(test_default)
-    pe.PROJECT_CONFIG_FILE.write_text(
-        pyaud.config.toml.dumps(rcfile), pe.ENCODING
-    )
-    pyaud.config.load_config()
-    pyaud.config.configure_logging()
+    pe.PROJECT_CONFIG_FILE.write_text(pc.toml.dumps(rcfile), pe.ENCODING)
+    pc.load_config()
+    pc.configure_logging()
     logger = logging.getLogger("default").root
     handler = logger.handlers[0]
     assert isinstance(handler, logging_handlers.TimedRotatingFileHandler)
@@ -723,11 +712,9 @@ def test_filter_logging_config_kwargs() -> None:
 
     # patch `DEFAULT_CONFIG` for `StreamHandler`
     rcfile["logging"]["handlers"]["default"]["class"] = "logging.StreamHandler"
-    pe.PROJECT_CONFIG_FILE.write_text(
-        pyaud.config.toml.dumps(rcfile), pe.ENCODING
-    )
-    pyaud.config.load_config()
-    pyaud.config.configure_logging()
+    pe.PROJECT_CONFIG_FILE.write_text(pc.toml.dumps(rcfile), pe.ENCODING)
+    pc.load_config()
+    pc.configure_logging()
     logger = logging.getLogger("default").root
     handler = logger.handlers[0]
     assert isinstance(handler, logging.StreamHandler)
@@ -738,7 +725,7 @@ def test_filter_logging_config_kwargs() -> None:
 def test_default_key() -> None:
     """Test setting and restoring of existing dict keys."""
     obj = {"default_key": "default_value"}
-    with pyaud.config.TempEnvVar(  # pylint: disable=protected-access
+    with pc.TempEnvVar(  # pylint: disable=protected-access
         obj, default_key="temp_value"
     ):
         assert obj["default_key"] == "temp_value"
@@ -831,8 +818,10 @@ def test_get_commit_hash(
     :param returncode: Mock return code from subprocess.
     :param expected: Expected result.
     """
-    monkeypatch.setattr("pyaud.git.rev_parse", lambda *_, **__: returncode)
-    monkeypatch.setattr("pyaud.git.stdout", lambda: stdout)
+    monkeypatch.setattr(
+        "pyaud._utils.git.rev_parse", lambda *_, **__: returncode
+    )
+    monkeypatch.setattr("pyaud._utils.git.stdout", lambda: stdout)
     assert pyaud._utils.get_commit_hash() == expected
 
 
@@ -917,13 +906,12 @@ def test_nested_times(monkeypatch: pytest.MonkeyPatch, main: t.Any) -> None:
             "<class 'tests._test.test_nested_times.<locals>.P2'>": [1],
         }
     }
-    default_config = pyaud.config.DEFAULT_CONFIG
+    default_config = pc.DEFAULT_CONFIG
     test_default: t.Dict[t.Any, t.Any] = copy.deepcopy(default_config)
     test_default["audit"]["modules"] = ["plugin_1", "plugin_2"]
-    pe.GLOBAL_CONFIG_FILE.write_text(
-        pyaud.config.toml.dumps(test_default), pe.ENCODING
-    )
+    pe.GLOBAL_CONFIG_FILE.write_text(pc.toml.dumps(test_default), pe.ENCODING)
 
+    # noinspection PyUnresolvedReferences
     pyaud.plugins.register("audit")(pyaud._default._Audit)
 
     class P1(pyaud.plugins.Action):
@@ -967,64 +955,29 @@ def test_del_key_config_runtime(main: t.Any) -> None:
             """Nothing to do."""
 
     # check config file for essential key
-    pyaud.config.toml.loads(pe.GLOBAL_CONFIG_FILE.read_text(pe.ENCODING))
+    pc.toml.loads(pe.GLOBAL_CONFIG_FILE.read_text(pe.ENCODING))
 
-    assert "filename" in pyaud.config.toml["logging"]["handlers"]["default"]
+    assert "filename" in pc.toml["logging"]["handlers"]["default"]
 
-    del pyaud.config.toml["logging"]["handlers"]["default"]["filename"]
+    del pc.toml["logging"]["handlers"]["default"]["filename"]
 
-    pe.GLOBAL_CONFIG_FILE.write_text(pyaud.config.toml.dumps(), pe.ENCODING)
+    pe.GLOBAL_CONFIG_FILE.write_text(pc.toml.dumps(), pe.ENCODING)
 
     # check config file to confirm essential key was removed
-    pyaud.config.toml.loads(pe.GLOBAL_CONFIG_FILE.read_text(pe.ENCODING))
+    pc.toml.loads(pe.GLOBAL_CONFIG_FILE.read_text(pe.ENCODING))
 
-    assert (
-        "filename" not in pyaud.config.toml["logging"]["handlers"]["default"]
-    )
+    assert "filename" not in pc.toml["logging"]["handlers"]["default"]
 
-    pe.GLOBAL_CONFIG_FILE.write_text(pyaud.config.toml.dumps(), pe.ENCODING)
+    pe.GLOBAL_CONFIG_FILE.write_text(pc.toml.dumps(), pe.ENCODING)
 
-    pyaud.config.configure_global()
+    pc.configure_global()
     main("plugin")
 
     # confirm after running main that no crash occurred and that the
     # essential key was replaced with a default
-    pyaud.config.toml.loads(pe.GLOBAL_CONFIG_FILE.read_text(pe.ENCODING))
+    pc.toml.loads(pe.GLOBAL_CONFIG_FILE.read_text(pe.ENCODING))
 
-    assert "filename" in pyaud.config.toml["logging"]["handlers"]["default"]
-
-
-@pytest.mark.parametrize("temp,expected", [(True, False), (False, True)])
-def test_call_m2r_on_markdown(
-    monkeypatch: pytest.MonkeyPatch, temp: bool, expected: bool
-) -> None:
-    """Test creation of an RST README when only markdown is present.
-
-    :param monkeypatch: Mock patch environment and attributes.
-    :param temp: Is the RST file temporary? True or False.
-    :param expected: Expected value of ``Path(...).is_file``.
-    """
-
-    # noinspection PyUnusedLocal
-    @pyaud.plugins.register(name="plugin")
-    class Plugin(pyaud.plugins.Action):
-        """Nothing to do."""
-
-        def action(self, *args: t.Any, **kwargs: bool) -> t.Any:
-            """Nothing to do."""
-
-    old_path = Path.cwd() / "README.md"
-    new_path = Path.cwd() / "README.rst"
-    old_path.touch()
-    tracker = Tracker()
-    tracker.return_values.append("rst text")
-    monkeypatch.setattr("pyaud.parsers._m2r.parse_from_file", tracker)
-    with pyaud.parsers.Md2Rst(old_path, temp=temp):
-        # do stuff here
-        pass
-
-    assert tracker
-    assert new_path.is_file() == expected
+    assert "filename" in pc.toml["logging"]["handlers"]["default"]
 
 
 def test_command_not_found_error() -> None:
@@ -1146,27 +1099,6 @@ def test_audit_error_did_no_pass_all_checks(
         main("plugin")
 
 
-def test_readme_replace() -> None:
-    """Test that ``LineSwitch`` properly edits a file."""
-    path = Path.cwd() / README
-
-    def _test_file_index(title: str, underline: str) -> None:
-        lines = path.read_text(pe.ENCODING).splitlines()
-        assert lines[0] == title
-        assert lines[1] == len(underline) * "="
-
-    repo = "repo"
-    readme = "README"
-    repo_underline = len(repo) * "="
-    readme_underline = len(readme) * "="
-    path.write_text(f"{repo}\n{repo_underline}\n", pe.ENCODING)
-    _test_file_index(repo, repo_underline)
-    with pyaud.parsers.LineSwitch(path, {0: readme, 1: readme_underline}):
-        _test_file_index(readme, readme_underline)
-
-    _test_file_index(repo, repo_underline)
-
-
 def test_no_exe_provided(monkeypatch: pytest.MonkeyPatch) -> None:
     """Test default value for exe property.
 
@@ -1216,9 +1148,9 @@ def test_clean_exclude(
     :param expected: Expected output from ``pyaud clean``.
     """
     Path(Path.cwd() / README).touch()
-    pyaud.git.init(devnull=True)  # type: ignore
-    pyaud.git.add(".")  # type: ignore
-    pyaud.git.commit("-m", "Initial commit", devnull=True)  # type: ignore
+    git.init(devnull=True)  # type: ignore
+    git.add(".")  # type: ignore
+    git.commit("-m", "Initial commit", devnull=True)  # type: ignore
     for exclusion in exclude:
         Path(Path.cwd() / exclusion).touch()
 
@@ -1233,11 +1165,11 @@ def test_make_generate_rcfile(nocolorcapsys: NoColorCapsys) -> None:
         color codes.
     """
     # noinspection PyUnresolvedReferences
-    pyaud.register_default_plugins()  # type: ignore
+    pyaud._default.register_default_plugins()  # type: ignore
     pyaud.plugins.get("generate-rcfile")()
     assert (
         nocolorcapsys.stdout().strip()
-        == pyaud.config.toml.dumps(pyaud.config.DEFAULT_CONFIG).strip()
+        == pc.toml.dumps(pc.DEFAULT_CONFIG).strip()
     )
 
 
@@ -1273,7 +1205,7 @@ def test_audit_modules(
     :param add: Function to add to the ``audit_modules`` list
     :param first: Expected first function executed.
     """
-    seq = list(pyaud.config.DEFAULT_CONFIG["audit"]["modules"])
+    seq = list(pc.DEFAULT_CONFIG["audit"]["modules"])
     seq.extend(add)
     mapping = {i: call_status(i) for i in seq}
     monkeypatch.setattr(PYAUD_PLUGINS_PLUGINS, mapping)
@@ -1341,12 +1273,10 @@ def test_suppress(
     :param monkeypatch: Mock patch environment and attributes.
     :param make_tree: Create directory tree from dict mapping.
     """
-    default_config = pyaud.config.DEFAULT_CONFIG
+    default_config = pc.DEFAULT_CONFIG
     test_default: t.Dict[t.Any, t.Any] = copy.deepcopy(default_config)
     test_default["audit"]["modules"] = ["plugin"]
-    pe.GLOBAL_CONFIG_FILE.write_text(
-        pyaud.config.toml.dumps(test_default), pe.ENCODING
-    )
+    pe.GLOBAL_CONFIG_FILE.write_text(pc.toml.dumps(test_default), pe.ENCODING)
     make_tree(Path.cwd(), {FILES: None, "docs": {CONFPY: None}})
     pyaud.files.append(Path.cwd() / FILES)
     monkeypatch.setattr(SP_OPEN_PROC, lambda *_, **__: 1)